+++ relayer-cli/src/commands/tx/transfer.rs
@@ -1,15 +1,20 @@
-use crate::prelude::*;
 use std::sync::{Arc, Mutex};
-use tokio::runtime::Runtime as TokioRuntime;
 
 use abscissa_core::{Command, Options, Runnable};
-use relayer::config::Config;
+use serde_json::json;
+use tokio::runtime::Runtime as TokioRuntime;
 
-use crate::error::{Error, Kind};
 use ibc::events::IBCEvent;
 use ibc::ics24_host::identifier::{ChannelId, PortId};
-use relayer::chain::{Chain, CosmosSDKChain};
-use relayer::transfer::{build_and_send_transfer_messages, TransferOptions};
+use relayer::{
+    chain::{Chain, CosmosSDKChain},
+    config::Config,
+    transfer::{build_and_send_transfer_messages, TransferOptions},
+};
+
+use crate::conclude::Output;
+use crate::error::{Error, Kind};
+use crate::prelude::*;
 
 #[derive(Clone, Command, Debug, Options)]
 pub struct TxRawSendPacketCmd {
@@ -69,8 +74,7 @@ impl Runnable for TxRawSendPacketCmd {
 
         let opts = match self.validate_options(&config) {
             Err(err) => {
-                status_err!("invalid options: {}", err);
-                return;
+                return Output::with_error().with_result(json!(err)).exit();
             }
             Ok(result) => result,
         };
@@ -87,12 +91,10 @@ impl Runnable for TxRawSendPacketCmd {
                 .map_err(|e| Kind::Tx.context(e).into());
 
         match res {
-            Ok(ev) => status_info!(
-                "packet recv, result: ",
-                "{:#?}",
-                serde_json::to_string(&ev).unwrap()
-            ),
-            Err(e) => status_info!("packet recv failed, error: ", "{}", e),
+            Ok(ev) => Output::with_success().with_result(json!(ev)).exit(),
+            Err(e) => Output::with_error()
+                .with_result(json!(format!("{}", e)))
+                .exit(),
         }
     }
 }